requirements.txt
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8